core: harden cgroups-agent forwarding
On dbus1, we receive systemd1.Agent signals via the private socket, hence
it's trusted. However, on kdbus we receive it on the system bus. We must
make sure it's sent by UID=0, otherwise unprivileged users can fake it.

Furthermore, never forward broadcasts we sent ourself. This might happen
on kdbus, as we forward the message on the same bus we received it on,
thus ending up in an endless loop.

Conflicts:
	src/core/dbus.c

Change-Id: Ie2bff9b2dbb1689098aefa1b753b60d7a073e1dd
Backported-by: Karol Lewandowski <k.lewandowsk@samsung.com>
Signed-off-by: Karol Lewandowski <k.lewandowsk@samsung.com>