busctl: add usage of Monitoring interface for non-kernel connections
This commit is loosely based on Lars Uebernickel's commit
(reverted in 8f2f0ffe8246) in a way that it takes
his implementation of calling the Monitoring interface.
Unfortunately the original commit broke kernel connections. That's why
this commit makes it different for kernel and non-kernel connections.

To make it cleaner, preparing connections is extracted to separate
functions.

From Lars Uebernickel's original commit message - about Monitoring interface:
This is now the recommended way to do monitoring by upstream D-Bus.
It's also allowed in the default policy, whereas eavesdrop is not
anymore, which effectively broke busctl on many systems.

Change-Id: I886fff5a3ec49df6aab26b217d881f249591ca9a